backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
python-decouple>=3.8
Pillow>=10.0.0
